Remove output device from LEDOutput
Since LED output configuration can be different for every setup, storing the output device in the LEDOutput effect can lead to problems and required a hack to override the output device.
This hack didn't always work (see #68)
The configuration of the output device is now part of the application.

Fixes #68.